move core to src folder, init gitignore file
